% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctx.R
\name{init_zstd_dctx}
\alias{init_zstd_dctx}
\title{Initialise a ZSTD compression context}
\usage{
init_zstd_dctx(dict = NULL)
}
\arguments{
\item{dict}{Dictionary. Default: NULL (no dictioary)}
}
\value{
ZSTD Decompression Context
}
\description{
Initialise a ZSTD compression context
}
