% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctx.R
\name{zstd_cctx_settings}
\alias{zstd_cctx_settings}
\title{Get the configuration settings of a compression context}
\usage{
zstd_cctx_settings(cctx)
}
\arguments{
\item{cctx}{ZSTD compression context, as created by \code{zstd_cctx()}}
}
\value{
named list of configuration options
}
\description{
Get the configuration settings of a compression context
}
\examples{
cctx <- zstd_cctx()
zstd_cctx_settings(cctx)
}
