% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctx.R
\name{zstd_dctx_settings}
\alias{zstd_dctx_settings}
\title{Get the configuration settings of a decompression context}
\usage{
zstd_dctx_settings(dctx)
}
\arguments{
\item{dctx}{ZSTD decompression context, as created by \code{zstd_dctx()}}
}
\value{
named list of configuration options
}
\description{
Get the configuration settings of a decompression context
}
\examples{
dctx <- zstd_dctx()
zstd_dctx_settings(dctx)
}
