% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{zstd_dict_id}
\alias{zstd_dict_id}
\title{Get the ID of a dictionary}
\usage{
zstd_dict_id(dict)
}
\arguments{
\item{dict}{raw vector}
}
\value{
dictionary id.  if data does not represent a dictionary, returns 0
}
\description{
Get the ID of a dictionary
}
