% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstd-info.R
\name{zstd_info}
\alias{zstd_info}
\title{Return information about the zstd stream}
\usage{
zstd_info(src)
}
\arguments{
\item{src}{raw vector, file or connection}
}
\value{
named list with \code{compressed_size}, \code{uncompressed_size}, 
        \code{dict_id} and \code{has_checksum}.  If an error occurs, or 
        the data does not appear to represent Zstandard compressed data,
        function returns \code{NULL}
}
\description{
Return information about the zstd stream
}
\examples{
data <- as.raw(sample(1:2, 10000, replace = TRUE))
cdata <- zstd_compress(data)
zstd_info(cdata)
}
