% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize-stream.R
\name{zstd_serialize_stream}
\alias{zstd_serialize_stream}
\title{Serialize arbitrary objects to a compressed stream of bytes using Zstandard}
\usage{
zstd_serialize_stream(robj, level = 3L, num_threads = 1L)
}
\arguments{
\item{robj}{Any R object understood by \code{base::serialize()}}

\item{level}{compression level -5 to 22. Default: 3}

\item{num_threads}{number of threads to use for compression}
}
\value{
serialized representation compressed into a raw byte vector
}
\description{
Serialize arbitrary objects to a compressed stream of bytes using Zstandard
}
