% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{zstd_train_dict_serialize}
\alias{zstd_train_dict_serialize}
\title{Train a dictionary}
\usage{
zstd_train_dict_serialize(
  samples,
  size = 1e+05,
  optim = FALSE,
  optim_shrink_allow = 0
)
}
\arguments{
\item{samples}{list of example R objects to train a dictionary to be 
used with \code{zstd_serialize}}

\item{size}{size of dictionary in bytes. Recommend: 100,000}

\item{optim}{optimize the dictionary. Default FALSE.  If TRUE, then ZSTD
will spend time optimizing the dictionary.}

\item{optim_shrink_allow}{integer value representing a percentage.
If non-zero, then a search will be carried out for dictionaries of a 
smaller size which are up to optim_shrink_allow percent worse than
the maximum sized dictionary.  Default: 0 means that no 
shrinking will be done.}
}
\value{
raw vector containing a ZSTD dictionary
}
\description{
Train a dictionary
}
