% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize-stream.R
\name{zstd_unserialize_stream}
\alias{zstd_unserialize_stream}
\title{Unserialize stream}
\usage{
zstd_unserialize_stream(src, dctx = NULL)
}
\arguments{
\item{src}{raw vector}

\item{dctx}{context}
}
\description{
Unserialize stream
}
