% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize-stream.R
\name{zstd_unserialize_stream_file}
\alias{zstd_unserialize_stream_file}
\title{Unserialize stream from file}
\usage{
zstd_unserialize_stream_file(src, dctx = NULL)
}
\arguments{
\item{src}{raw vector}

\item{dctx}{context}
}
\description{
Unserialize stream from file
}
