% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{zstd_version}
\alias{zstd_version}
\title{Get version string of zstd C library}
\usage{
zstd_version()
}
\value{
String containing version number of zstd C library
}
\description{
Get version string of zstd C library
}
\examples{
zstd_version()
}
