% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstdfile.R
\name{zstdfile}
\alias{zstdfile}
\title{Create a file connection which uses Zstandard compression.}
\usage{
zstdfile(description, open = "", ..., cctx = NULL, dctx = NULL)
}
\arguments{
\item{description}{zstandard filename}

\item{open}{character string. A description of how to open the connection if 
it is to be opened upon creation e.g. "rb". Default "" (empty string) means
to not open the connection on creation - user must still call \code{open()}.
Note: If an "open" string is provided, the user must still call \code{close()}
otherwise the contents of the file aren't completely flushed until the
connection is garbage collected.}

\item{...}{Other named arguments which override the contexts e.g. \code{level = 20}}

\item{cctx, dctx}{compression/decompression contexts created by 
\code{zstd_cctx()} and \code{zstd_dctx()}. Optional.}
}
\description{
Create a file connection which uses Zstandard compression.
}
\details{
This \code{zstdfile()} connection works like R's built-in connections (e.g.
\code{gzfile()}, \code{xzfile()}) but using the Zstandard algorithm
to compress/decompress the data.

This connection works with both ASCII and binary data, e.g. using 
\code{readLines()} and \code{readBin()}.
}
\examples{
tmp <- tempfile()
dat <- as.raw(1:255)
writeBin(dat, zstdfile(tmp, level = 20))
readBin(zstdfile(tmp),  raw(), 1000)
txt <- as.character(mtcars)
writeLines(txt, zstdfile(tmp))
readLines(zstdfile(tmp))
}
