% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctx.R
\name{init_zstd_cctx}
\alias{init_zstd_cctx}
\title{Initialise a ZSTD compression context}
\usage{
init_zstd_cctx(level = 3L, num_threads = 1L)
}
\arguments{
\item{level}{compression level. Default: 3}

\item{num_threads}{number of compression threads. Default 1}
}
\value{
ZSTD context
}
\description{
Initialise a ZSTD compression context
}
