% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{zstd_compress}
\alias{zstd_compress}
\title{Serialize arbitrary objects to a compressed stream of bytes using Zstandard}
\usage{
zstd_compress(raw_vec, level = 3L)
}
\arguments{
\item{raw_vec}{raw vector}

\item{level}{compression level -5 to 22. Default: 3}
}
\value{
raw vector
}
\description{
Serialize arbitrary objects to a compressed stream of bytes using Zstandard
}
