% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{zstd_compress}
\alias{zstd_compress}
\title{Serialize arbitrary objects to a compressed stream of bytes using Zstandard}
\usage{
zstd_compress(src, file = NULL, ..., cctx = NULL)
}
\arguments{
\item{src}{Source data to be compressed.  This may be a raw vector, or a
character string}

\item{file}{filename in which to serialize data. If NULL (the default), then 
serialize the results to a raw vector}

\item{...}{extra arguments passed to context initializer}

\item{cctx}{ZSTD Compression Context created by \code{init_zstd_cctx()} or NULL.
Default: NULL will create a context on the fly}
}
\value{
raw vector of compressed data
}
\description{
Serialize arbitrary objects to a compressed stream of bytes using Zstandard
}
