% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{zstd_decompress}
\alias{zstd_decompress}
\title{Decompress raw bytes}
\usage{
zstd_decompress(raw_vec, dctx = NULL)
}
\arguments{
\item{raw_vec}{raw vector}

\item{dctx}{ZSTD Decompression Context}
}
\description{
Decompress raw bytes
}
