% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstd.R
\name{zstd_decompress}
\alias{zstd_decompress}
\title{Unpack compressed bytes into its original vector}
\usage{
zstd_decompress(raw_vec)
}
\arguments{
\item{raw_vec}{Vector of raw values representing the compressed data.
This data must have been created with \code{zstdlite::zstd_compress()}}
}
\description{
Unpack compressed bytes into its original vector
}
