% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{zstd_decompress}
\alias{zstd_decompress}
\title{Decompress raw bytes}
\usage{
zstd_decompress(src, type = "raw", ..., dctx = NULL)
}
\arguments{
\item{src}{Raw vector or filename containing a ZSTD compressed serialized representation of
an R object}

\item{type}{should data be returned as a 'raw' vector? or as a 'string'? 
Default: 'raw'}

\item{...}{extra arguments passed to context initializer}

\item{dctx}{ZSTD Decompression Context}
}
\description{
Decompress raw bytes
}
