% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{zstd_decompress}
\alias{zstd_decompress}
\title{Decompress raw bytes}
\usage{
zstd_decompress(
  src,
  type = "raw",
  ...,
  dctx = NULL,
  use_file_streaming = FALSE
)
}
\arguments{
\item{src}{Raw vector or filename containing a ZSTD compressed serialized representation of
an R object}

\item{type}{Should data be returned as a 'raw' vector or as a 'string'? 
Default: 'raw'}

\item{...}{extra arguments passed to \code{zstd_cctx()} or \code{zstd_dctx()}
context initializers. 
Note: These argument are only used when \code{cctx} or \code{dctx} is NULL}

\item{dctx}{ZSTD Decompression Context created by \code{zstd_dctx()} or NULL.
Default: NULL will create a default decompression context on-the-fly.}

\item{use_file_streaming}{Use the streaming interface when reading or writing
to a file?  This may reduce memory allocations
and make better use of mutlithreading.  Default: FALSE}
}
\description{
Decompress raw bytes
}
