% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{zstd_dict_id}
\alias{zstd_dict_id}
\title{Get the Dictionary ID of a dictionary or compressed data}
\usage{
zstd_dict_id(dict)
}
\arguments{
\item{dict}{raw vector or filename}
}
\value{
Signed integer value representing the Dictionary ID. If data does not 
        represent a dictionary, or data which was compressed with a dictionary,
        then a value of 0 is returned.
}
\description{
Get the Dictionary ID of a dictionary or compressed data
}
