% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{zstd_serialize}
\alias{zstd_serialize}
\alias{zstd_unserialize}
\title{Serialize arbitrary objects to a compressed stream of bytes using Zstandard}
\usage{
zstd_serialize(robj, file = NULL, ..., cctx = NULL)

zstd_unserialize(src, ..., dctx = NULL)
}
\arguments{
\item{robj}{Any R object understood by \code{base::serialize()}}

\item{file}{filename in which to serialize data. If NULL (the default), then 
serialize the results to a raw vector}

\item{...}{extra arguments passed to context initializer}

\item{cctx}{ZSTD Compression Context created by \code{init_zstd_cctx()} or NULL.
Default: NULL will create a context on the fly}

\item{src}{Raw vector or filename containing a ZSTD compressed serialized representation of
an R object}

\item{dctx}{ZSTD Decompression Context created by \code{init_zstd_dctx()} or NULL.
Default: NULL will decompression without any context using default seetings.}
}
\value{
serialized representation compressed into a raw byte vector
}
\description{
Serialize arbitrary objects to a compressed stream of bytes using Zstandard
}
